+++ R/auk-species.r
@@ -36,7 +36,7 @@ auk_species.auk_ebd <- function(x, species, replace = FALSE) {
     is.character(species),
     assertthat::is.flag(replace)
   )
-  species_clean <- ebird_species(species)
+  species_clean <- ebird_species(species, type = "scientific")
 
   # check all species names are valid
   if (any(is.na(species_clean))) {